+++ CHANGELOG.md
@@ -70,3 +70,8 @@ All notable changes to this project will be documented in this file.
 ## [1.2.4]
 ### Fixes
 - header link color
+
+
+## [1.2.5]
+### Fixes
+- fix template_include for elementor